requirements.txt
# dlt: BigQuery + filesystem (GCS) destinations
dlt[bigquery]>=1.0.0
dlt[filesystem]>=1.0.0
dlt[gs]>=1.0.0

# Data & APIs
pandas>=2.0.0
pyarrow>=14.0.0
datasets>=2.14.0
kaggle>=1.5.0

# Validation
pydantic>=2.0.0

# Skills extraction (LLM)
google-generativeai>=0.8.0
